fixing alpha and zero default directories
