+++ dgraph/cmd/alpha/run.go
@@ -675,6 +675,7 @@ func run() {
 		TLSClientConfig:      tlsClientConf,
 		TLSServerConfig:      tlsServerConf,
 		HmacSecret:           opts.HmacSecret,
+		AuditDir:             Alpha.Conf.GetString("audit_dir"),
 	}
 	x.WorkerConfig.Parse(Alpha.Conf)
 	x.CheckFlag(x.WorkerConfig.Raft, "group", "idx", "learner")
